+++ src/Connector/RecupPesS2low/RecupPesS2lowConnector.php
@@ -2,12 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Pastell\Connector\RecupActesS2low;
+namespace Pastell\Connector\RecupPesS2low;
 
 use DonneesFormulaire;
-use Pastell\Actes\ActeEnvelopeParser;
-use Pastell\Client\S2low\Model\ActeListQuery;
-use Pastell\Client\S2low\Model\ActeListResponse;
+use Pastell\Client\S2low\Model\PesAllerListQuery;
+use Pastell\Client\S2low\Model\PesAllerListResponse;
 use Pastell\Client\S2low\S2lowClient;
 use Pastell\Client\S2low\S2lowClientAuth;
 use Pastell\Client\S2low\S2lowClientException;
@@ -16,11 +15,11 @@ use Pastell\Service\Document\DocumentDeletionService;
 use Psr\Http\Client\ClientExceptionInterface;
 use Recuperateur;
 
-class RecupActesS2lowConnector extends \Connecteur
+class RecupPesS2lowConnector extends \Connecteur
 {
-    private const FLUX = 'ls-recup-actes-s2low';
-    private const STATUS_ACK = 4;
-    private const STATUS_SENT_TO_SAE = '12';
+    private const FLUX = 'ls-recup-pes-s2low';
+    private const STATUS_AVAILABLE = 8;
+    private const STATUS_SENT_TO_SAE = '19';
     private S2lowClient $client;
     private string $startDate;
     private string $endDate;
@@ -40,26 +39,14 @@ class RecupActesS2lowConnector extends \Connecteur
     }
 
     /**
-     * @throws \DateInvalidOperationException
-     * @throws \DateMalformedStringException
      * @throws \UnrecoverableException
      */
     public function setConnecteurConfig(DonneesFormulaire $donneesFormulaire): void
     {
         $url = $donneesFormulaire->get('url');
-        $this->transactionStatus = (int)$donneesFormulaire->get('transaction_status') ?: self::STATUS_ACK;
+        $this->transactionStatus = (int)$donneesFormulaire->get('transaction_status') ?: self::STATUS_AVAILABLE;
         $this->startDate = $donneesFormulaire->get('start_date');
         $this->endDate = $donneesFormulaire->get('end_date');
-
-        $dateSixtyDaysAgo = new \DateTime();
-        $dateSixtyDaysAgo->sub(new \DateInterval('P62D'));
-        $connectorDate = new \DateTime($donneesFormulaire->get('end_date'));
-
-        if ($connectorDate > $dateSixtyDaysAgo  && $this->transactionStatus === self::STATUS_ACK) {
-            $this->endDate = $dateSixtyDaysAgo->format('Y-m-d');
-        } else {
-            $this->endDate = $connectorDate->format('Y-m-d');
-        }
 
         $this->numberOfDocumentsPerJob = (int)$donneesFormulaire->get('nb_recup') ?: 10;
         $this->maxNumberOfDocumentsInEntity = (int)$donneesFormulaire->get('nb_documents') ?: 100;
@@ -67,7 +54,6 @@ class RecupActesS2lowConnector extends \Connecteur
         $auth = new S2lowClientAuth();
         $auth->username = $donneesFormulaire->get('user_login') ?: '';
         $auth->password = $donneesFormulaire->get('user_password') ?: '';
-
         $auth->user_certificat_password = $donneesFormulaire->get('certificate_password');
         $auth->user_key_pem = $donneesFormulaire->getFilePath('certificate_key');
         $auth->user_certificat_pem = $donneesFormulaire->getFilePath('certificate_pem');
@@ -112,18 +98,17 @@ class RecupActesS2lowConnector extends \Connecteur
     /**
      * @throws S2lowClientException
      * @throws ClientExceptionInterface
-     * @throws \JsonException
      */
-    public function listActes(int $numberOfTransactions, int $offset = 0): ActeListResponse
+    public function listPes(int $numberOfTransactions, int $offset = 0): PesAllerListResponse
     {
-        $query = new ActeListQuery();
+        $query = new PesAllerListQuery();
         $query->limit = $numberOfTransactions;
         $query->minDate = $this->startDate;
         $query->maxDate = $this->endDate;
         $query->offset = $offset;
         $query->statusId = $this->transactionStatus;
 
-        return $this->client->actes()->getActesList($query);
+        return $this->client->pes()->getPesAllerList($query);
     }
 
     /**
@@ -132,7 +117,7 @@ class RecupActesS2lowConnector extends \Connecteur
      */
     public function changeStatus(string $transactionId, string $statusId): void
     {
-        $this->client->actes()->changeActeStatus($transactionId, $statusId);
+        $this->client->pes()->changePesStatus($transactionId, $statusId);
     }
 
     /**
@@ -140,9 +125,8 @@ class RecupActesS2lowConnector extends \Connecteur
      * @throws ClientExceptionInterface
      * @throws \NotFoundException
      * @throws S2lowClientException
-     * @throws \JsonException
      */
-    public function fetchActes(): array
+    public function fetchPesAller(): array
     {
         $entityId = $this->getConnecteurInfo()['id_e'];
         $createdDocuments = 0;
@@ -150,27 +134,24 @@ class RecupActesS2lowConnector extends \Connecteur
         $numberOfDocumentsToCreate = $this->getNumberOfDocumentsToCreate($entityId);
         $message = [];
         while ($createdDocuments < $numberOfDocumentsToCreate) {
-            $listActes = $this->listActes($numberOfDocumentsToCreate, $offset);
-            if (count($listActes->transactions) === 0) {
+            $listPes = $this->listPes($numberOfDocumentsToCreate, $offset);
+            if (count($listPes->transactions) === 0) {
                 break;
             }
 
-            foreach ($listActes->transactions as $transaction) {
-                if (!$transaction->isActes()) {
-                    continue;
-                }
+            foreach ($listPes->transactions as $transaction) {
+                $transactionId = $transaction['id'];
                 $documentId = $this->documentCreationService->createDocumentWithoutAuthorizationChecking(
                     $entityId,
                     self::FLUX
                 );
                 try {
-                    $this->createDocument($entityId, $documentId, $transaction->id);
+                    $this->createDocument($entityId, $documentId, $transactionId);
                     ++$createdDocuments;
 
                     $message[] = \sprintf(
-                        'Création du document lié à la transaction %s %s : %s',
-                        $transaction->id,
-                        $transaction->number,
+                        'Création du document lié à la transaction %s : %s',
+                        $transactionId,
                         $documentId,
                     );
                     if ($createdDocuments >= $numberOfDocumentsToCreate) {
@@ -178,9 +159,8 @@ class RecupActesS2lowConnector extends \Connecteur
                     }
                 } catch (\Throwable $e) {
                     $message[] = \sprintf(
-                        'Impossible de créer le document lié à la transaction %s %s => %s',
-                        $transaction->id,
-                        $transaction->number,
+                        'Impossible de créer le document lié à la transaction %s => %s',
+                        $transactionId,
                         $e->getMessage(),
                     );
                     $this->documentDeletionService->delete($documentId);
@@ -200,47 +180,25 @@ class RecupActesS2lowConnector extends \Connecteur
      */
     private function createDocument(int $entityId, string $documentId, string $transactionId): void
     {
-        $transactionFiles = $this->client->actes()->getFileList($transactionId);
-        $numberOfFiles = \count($transactionFiles);
-        $xml = $this->client->actes()->downloadFile($transactionFiles[0]->id);
-        $envelopeParser = new ActeEnvelopeParser($xml);
-
         $form = $this->formFactory->get($documentId);
 
+        $pesContent = $this->client->pes()->getPes($transactionId);
         $form->addFileFromData(
-            'arrete',
-            $transactionFiles[1]->postedFilename,
-            $this->client->actes()->downloadFile($transactionFiles[1]->id)
+            'fichier_pes',
+            'pes_aller.xml',
+            $pesContent
         );
-        if ($numberOfFiles > 2) {
-            for ($i = 2; $i < $numberOfFiles; ++$i) {
-                $form->addFileFromData(
-                    'autre_document_attache',
-                    $transactionFiles[$i]->postedFilename,
-                    $this->client->actes()->downloadFile($transactionFiles[$i]->id),
-                    $i - 2
-                );
-            }
-        }
-
-        $aractes = $this->client->actes()->getAractes($transactionId);
         $form->addFileFromData(
-            'aractes',
-            'aractes.xml',
-            $aractes
+            'pes_acquit',
+            'pes_acquit.xml',
+            $this->client->pes()->getPesAcquit($transactionId)
         );
-        $aractesParser = new ActeEnvelopeParser($aractes);
 
+        $pesAllerFile = new \PESAllerFile();
         $recuperateur = new Recuperateur([
             'transaction_id' => $transactionId,
-            'acte_nature' => $envelopeParser->getCodeNatureActe(),
-            'numero_de_lacte' => $envelopeParser->getNumeroInterne(),
-            'objet' => $envelopeParser->getObjet(),
-            'date_de_lacte' => $envelopeParser->getDate(),
-            'document_papier' => $envelopeParser->hasDocumentPapier(),
-            'classification' => $envelopeParser->getClassification(),
-            'date_ar' => $aractesParser->getDateAr(),
-            'update_status_actes_s2low_status_1' => self::STATUS_SENT_TO_SAE,
+            'objet' => $pesAllerFile->getAllInfo($pesContent, false)[\PESAllerFile::NOM_FIC],
+            'update_status_pes_s2low_status_1' => self::STATUS_SENT_TO_SAE,
         ]);
         $this->documentModificationService->modifyDocumentWithoutAuthorizationChecking(
             $entityId,