+++ src/Service/Connecteur/ConnecteurCreationService.php
@@ -3,9 +3,11 @@
 namespace Pastell\Service\Connecteur;
 
 use ConnecteurEntiteSQL;
+use ConnecteurException;
 use ConnecteurFactory;
 use DonneesFormulaireFactory;
 use Exception;
+use Pastell\Service\FeatureToggle\DisplayConnecteurEntiteRacine;
 
 class ConnecteurCreationService
 {
@@ -15,12 +17,14 @@ class ConnecteurCreationService
         private readonly ConnecteurActionService $connecteurActionService,
         private readonly ConnecteurAssociationService $connecteurAssociationService,
         private readonly DonneesFormulaireFactory $donneesFormulaireFactory,
+        private readonly DisplayConnecteurEntiteRacine $displayConnecteurEntiteRacine,
     ) {
     }
 
     /**
      * @param int<0,1> $global
      * @throws Exception
+     * @throws ConnecteurException
      */
     public function createConnecteur(
         string $connecteur_id,
@@ -33,6 +37,7 @@ class ConnecteurCreationService
         string $message = ''
     ): int {
 
+        $this->checkCreateConnector($id_e, $global);
         $libelle = ($libelle === '') ? $connecteur_id : $libelle;
 
         $id_ce =  $this->connecteurEntiteSQL->addConnecteur(
@@ -112,6 +117,24 @@ class ConnecteurCreationService
             if ($field->getDefault()) {
                 $donneesFormulaire->setData($field->getName(), $field->getDefault());
             }
+        }
+    }
+
+    /**
+     * @throws ConnecteurException
+     */
+    private function checkCreateConnector(int $id_e, int $global): void
+    {
+        /** @deprecated 4.1.7, to be removed in v5 */
+        if (($id_e === 0) && ($global === 0) && !($this->displayConnecteurEntiteRacine->isEnabled())) {
+            throw new ConnecteurException(
+                "Il n'est pas possible de créer un connecteur d'entité au niveau de l'entité racine"
+            );
+        }
+        if (($id_e !== 0) && ($global === 1)) {
+            throw new ConnecteurException(
+                "Il n'est pas possible de créer un connecteur global au niveau d'une entité"
+            );
         }
     }
 }