+++ test/PHPUnit/PastellTestCase.php
@@ -127,6 +127,7 @@ abstract class PastellTestCase extends TestCase
             'suppl_recup_parapheur_corbeille_connector' => true,
             'suppl_recup_fin_parapheur' => true,
             'suppl_recup_actes_s2low' => true,
+            'suppl_recup_pes_s2low' => true,
             'module_connecteur_fast' => true,
             'module_connecteur_vitam' => true,
         ]);